+++ src/release_notes.md
@@ -11,7 +11,7 @@
 ### Microsoft.Azure.Functions.Worker.Extensions.Mcp <version>
 
 - Added structured content support for tools (#172)
-- Adding input schema generation in the worker and enabling `UseWorkerInputSchema` to be true (#166)
+- Generate the tool input schema in the worker (#166)
 
 ### Microsoft.Azure.Functions.Worker.Extensions.Mcp.Sdk <version>